Add requestId as a variable in the dpd-event context
This will add requestId as a variable for the dpd event resource scope,
it may be as an attribute, or header in the request object with
standard names.

The requestId allows to identify the same client through different
layers in the server for example when passing the same request through
multiple event resources and makes it possible to link the request and
do lookups based on the requestId.

dpd is lowercasing the headers names in the cors module (corser) so, if
is an external call, we get the request id with the normal name as an
attribute in the request due to Express Middleware, but if is an
internal call we get it through the headers.